+++ src/extensions/model-archive/quality-assessment/prop.ts
@@ -1,21 +1,24 @@
 /**
- * Copyright (c) 2021 mol* contributors, licensed under MIT, See LICENSE file for more info.
+ * Copyright (c) 2021-24 mol* contributors, licensed under MIT, See LICENSE file for more info.
  *
  * @author Alexander Rose <alexander.rose@weirdbyte.de>
  * @author David Sehnal <david.sehnal@gmail.com>
  */
 
-import { ParamDefinition as PD } from '../../../mol-util/param-definition';
-import { Unit } from '../../../mol-model/structure';
-import { CustomProperty } from '../../../mol-model-props/common/custom-property';
+import { CifFrame } from '../../../mol-io/reader/cif';
+import { toDatabase } from '../../../mol-io/reader/cif/schema';
+import { mmCIF_Schema } from '../../../mol-io/reader/cif/schema/mmcif';
+import { MmcifFormat } from '../../../mol-model-formats/structure/mmcif';
 import { CustomModelProperty } from '../../../mol-model-props/common/custom-model-property';
+import { CustomProperty } from '../../../mol-model-props/common/custom-property';
+import { CustomPropertyDescriptor } from '../../../mol-model/custom-property';
+import { Unit } from '../../../mol-model/structure';
 import { Model, ResidueIndex } from '../../../mol-model/structure/model';
-import { QuerySymbolRuntime } from '../../../mol-script/runtime/query/compiler';
+import { AtomicIndex } from '../../../mol-model/structure/model/properties/atomic';
 import { CustomPropSymbol } from '../../../mol-script/language/symbol';
 import { Type } from '../../../mol-script/language/type';
-import { CustomPropertyDescriptor } from '../../../mol-model/custom-property';
-import { MmcifFormat } from '../../../mol-model-formats/structure/mmcif';
-import { AtomicIndex } from '../../../mol-model/structure/model/properties/atomic';
+import { QuerySymbolRuntime } from '../../../mol-script/runtime/query/compiler';
+import { ParamDefinition as PD } from '../../../mol-util/param-definition';
 
 export { QualityAssessment };
 
@@ -26,10 +29,19 @@ interface QualityAssessment {
 }
 
 namespace QualityAssessment {
+    export interface Pairwise {
+        id: number
+        name: string
+
+        residueRange: [ResidueIndex, ResidueIndex]
+        valueRange: [number, number]
+        values: Record<ResidueIndex, Record<ResidueIndex, number | undefined> | undefined>
+    }
+
     const Empty = {
         value: {
-            localMetrics: new Map()
-        }
+            localMetrics: new Map(),
+        } satisfies QualityAssessment
     };
 
     export function isApplicable(model?: Model, localMetricName?: 'pLDDT' | 'qmean'): boolean {
@@ -104,6 +116,101 @@ namespace QualityAssessment {
                 qmean: localMetrics.get('qmean'),
             }
         };
+    }
+
+    const PairwiseSchema = {
+        ma_qa_metric: mmCIF_Schema.ma_qa_metric,
+        ma_qa_metric_local_pairwise: mmCIF_Schema.ma_qa_metric_local_pairwise
+    };
+
+    export function findModelArchiveCIFPAEMetrics(frame: CifFrame) {
+        const { ma_qa_metric, ma_qa_metric_local_pairwise } = toDatabase(PairwiseSchema, frame);
+        const result: { id: number, name: string }[] = [];
+        if (ma_qa_metric_local_pairwise._rowCount === 0) return result;
+
+        for (let i = 0, il = ma_qa_metric._rowCount; i < il; i++) {
+            if (ma_qa_metric.mode.value(i) !== 'local-pairwise') continue;
+            const id = ma_qa_metric.id.value(i);
+            const name = ma_qa_metric.name.value(i);
+            if (!name.toLowerCase().includes('pae')) continue;
+            result.push({ id, name });
+        }
+        return result;
+    }
+
+    export function pairwiseMetricFromModelArchiveCIF(model: Model, frame: CifFrame, metricId: number): Pairwise | undefined {
+        const db = toDatabase(PairwiseSchema, frame);
+        if (!db.ma_qa_metric_local_pairwise._rowCount) return undefined;
+
+        const { ma_qa_metric, ma_qa_metric_local_pairwise } = db;
+        const { model_id, label_asym_id_1, label_seq_id_1, label_asym_id_2, label_seq_id_2, metric_id, metric_value } = db.ma_qa_metric_local_pairwise;
+        const { index } = model.atomicHierarchy;
+
+        let metric: Pairwise | undefined;
+
+        for (let i = 0, il = ma_qa_metric._rowCount; i < il; i++) {
+            if (ma_qa_metric.mode.value(i) !== 'local-pairwise') continue;
+            const id = ma_qa_metric.id.value(i);
+            if (id !== metricId) continue;
+
+            const name = ma_qa_metric.name.value(i);
+            metric = {
+                id,
+                name,
+                residueRange: [Number.MAX_SAFE_INTEGER as ResidueIndex, Number.MIN_SAFE_INTEGER as ResidueIndex],
+                valueRange: [Number.MAX_VALUE, -Number.MAX_VALUE],
+                values: {}
+            };
+        }
+
+        if (!metric) return undefined;
+
+        const { values, residueRange, valueRange } = metric;
+        const residueKey: AtomicIndex.ResidueLabelKey = {
+            label_entity_id: '',
+            label_asym_id: '',
+            label_seq_id: 0,
+            pdbx_PDB_ins_code: undefined,
+        };
+
+        for (let i = 0, il = ma_qa_metric_local_pairwise._rowCount; i < il; i++) {
+            if (model_id.value(i) !== model.modelNum || metric_id.value(i) !== metricId) continue;
+
+            let labelAsymId = label_asym_id_1.value(i);
+            let entityIndex = index.findEntity(labelAsymId);
+            residueKey.label_entity_id = model.entities.data.id.value(entityIndex);
+            residueKey.label_asym_id = labelAsymId;
+            residueKey.label_seq_id = label_seq_id_1.value(i);
+
+            const rI_1 = index.findResidueLabel(residueKey);
+            if (rI_1 < 0) continue;
+
+            labelAsymId = label_asym_id_2.value(i);
+            entityIndex = index.findEntity(labelAsymId);
+            residueKey.label_entity_id = model.entities.data.id.value(entityIndex);
+            residueKey.label_asym_id = labelAsymId;
+            residueKey.label_seq_id = label_seq_id_2.value(i);
+
+            const rI_2 = index.findResidueLabel(residueKey);
+            if (rI_1 < 0) continue;
+
+            let r1 = values[rI_1];
+            if (!r1) {
+                r1 = {};
+                values[rI_1] = r1;
+            }
+            const value = metric_value.value(i);
+            r1[rI_2] = value;
+
+            if (rI_1 < residueRange[0]) residueRange[0] = rI_1;
+            if (rI_2 < residueRange[0]) residueRange[0] = rI_2;
+            if (rI_1 > residueRange[1]) residueRange[1] = rI_1;
+            if (rI_2 > residueRange[1]) residueRange[1] = rI_2;
+            if (value < valueRange[0]) valueRange[0] = value;
+            if (value > valueRange[1]) valueRange[1] = value;
+        }
+
+        return metric;
     }
 
     export const symbols = {