+++ src/mol-canvas3d/passes/draw.ts
@@ -47,6 +47,7 @@ export class DrawPass {
     private readonly drawTarget: RenderTarget;
 
     readonly colorTarget: RenderTarget;
+    readonly transparentColorTarget: RenderTarget;
     readonly depthTextureTransparent: Texture;
     readonly depthTextureOpaque: Texture;
 
@@ -91,6 +92,8 @@ export class DrawPass {
         const { extensions, resources, isWebGL2 } = webgl;
         this.drawTarget = createNullRenderTarget(webgl.gl);
         this.colorTarget = webgl.createRenderTarget(width, height, true, 'uint8', 'linear');
+        this.transparentColorTarget = webgl.createRenderTarget(width, height, false, 'uint8', 'linear');
+
         this.packedDepth = !extensions.depthTexture;
 
         this.depthTargetTransparent = webgl.createRenderTarget(width, height);
@@ -129,6 +132,7 @@ export class DrawPass {
         if (width !== w || height !== h) {
             this.colorTarget.setSize(width, height);
             this.depthTargetTransparent.setSize(width, height);
+            this.transparentColorTarget.setSize(width, height);
 
             if (this.depthTargetOpaque) {
                 this.depthTargetOpaque.setSize(width, height);
@@ -166,10 +170,9 @@ export class DrawPass {
             renderer.renderOpaque(scene.primitives, camera);
         }
 
-        const outlineEnabled = PostprocessingPass.isEnabled(postprocessingProps) && PostprocessingPass.isTransparentOutlineEnabled(postprocessingProps);
-        const dofEnabled = DofPass.isEnabled(postprocessingProps);
+        this.depthTextureOpaque.detachFramebuffer(this.colorTarget.framebuffer, 'depth');
 
-        if (outlineEnabled || dofEnabled) {
+        if (PostprocessingPass.isTransparentDepthRequired(scene, postprocessingProps)) {
             this.depthTargetTransparent.bind();
             renderer.clearDepth(true);
             if (scene.opacityAverage < 1) {
@@ -177,16 +180,14 @@ export class DrawPass {
             }
         }
 
-        if (PostprocessingPass.isEnabled(postprocessingProps)) {
-            this.postprocessing.render(camera, false, transparentBackground, renderer.props.backgroundColor, postprocessingProps, renderer.light, renderer.ambientColor);
-        }
-
-        this.depthTextureOpaque.detachFramebuffer(this.colorTarget.framebuffer, 'depth');
-
         // render transparent primitives
+        const isPostprocessingEnabled = PostprocessingPass.isEnabled(postprocessingProps);
         if (scene.opacityAverage < 1) {
-            const target = PostprocessingPass.isEnabled(postprocessingProps)
-                ? this.postprocessing.target : this.colorTarget;
+            const target = isPostprocessingEnabled ? this.transparentColorTarget : this.colorTarget;
+            if (isPostprocessingEnabled) {
+                target.bind();
+                renderer.clear(false, false, true);
+            }
 
             const dpoitTextures = this.dpoit.bind();
             renderer.renderDpoitTransparent(scene.primitives, camera, this.depthTextureOpaque, dpoitTextures);
@@ -206,6 +207,10 @@ export class DrawPass {
             this.dpoit.render();
         }
 
+        if (PostprocessingPass.isEnabled(postprocessingProps)) {
+            this.postprocessing.render(camera, scene, false, transparentBackground, renderer.props.backgroundColor, postprocessingProps, renderer.light, renderer.ambientColor);
+        }
+
         // render transparent volumes
         if (scene.volumes.renderables.length > 0) {
             renderer.renderVolume(scene.volumes, camera, this.depthTextureOpaque);
@@ -223,10 +228,7 @@ export class DrawPass {
             renderer.renderOpaque(scene.primitives, camera);
         }
 
-        const outlineEnabled = PostprocessingPass.isEnabled(postprocessingProps) && PostprocessingPass.isTransparentOutlineEnabled(postprocessingProps);
-        const dofEnabled = DofPass.isEnabled(postprocessingProps);
-
-        if (outlineEnabled || dofEnabled) {
+        if (PostprocessingPass.isTransparentDepthRequired(scene, postprocessingProps)) {
             this.depthTargetTransparent.bind();
             renderer.clearDepth(true);
             if (scene.opacityAverage < 1) {
@@ -234,28 +236,38 @@ export class DrawPass {
             }
         }
 
-        if (PostprocessingPass.isEnabled(postprocessingProps)) {
-            this.postprocessing.render(camera, false, transparentBackground, renderer.props.backgroundColor, postprocessingProps, renderer.light, renderer.ambientColor);
-        }
+        // render transparent primitives
+        const isPostprocessingEnabled = PostprocessingPass.isEnabled(postprocessingProps);
+        if (scene.opacityAverage < 1) {
+            const target = isPostprocessingEnabled ? this.transparentColorTarget : this.colorTarget;
+            if (isPostprocessingEnabled) {
+                target.bind();
+                renderer.clear(false, false, true);
+            }
 
-        // render transparent primitives and volumes
-        if (scene.opacityAverage < 1 || scene.volumes.renderables.length > 0) {
             this.wboit.bind();
-            if (scene.opacityAverage < 1) {
-                renderer.renderWboitTransparent(scene.primitives, camera, this.depthTextureOpaque);
-            }
-            if (scene.volumes.renderables.length > 0) {
-                renderer.renderWboitTransparent(scene.volumes, camera, this.depthTextureOpaque);
-            }
+            renderer.renderWboitTransparent(scene.primitives, camera, this.depthTextureOpaque);
 
             // evaluate wboit
-            if (PostprocessingPass.isEnabled(postprocessingProps)) {
-                this.postprocessing.target.bind();
-            } else {
-                this.colorTarget.bind();
-            }
+            target.bind();
             this.wboit.render();
         }
+
+        if (PostprocessingPass.isEnabled(postprocessingProps)) {
+            this.postprocessing.render(camera, scene, false, transparentBackground, renderer.props.backgroundColor, postprocessingProps, renderer.light, renderer.ambientColor);
+        }
+
+        // render volumes
+        if (scene.volumes.renderables.length > 0) {
+            this.wboit.bind();
+            renderer.renderWboitTransparent(scene.volumes, camera, this.depthTextureOpaque);
+
+            // evaluate wboit
+            const target = isPostprocessingEnabled ? this.postprocessing.target : this.colorTarget;
+            target.bind();
+            this.wboit.render();
+        }
+
     }
 
     private _renderBlended(renderer: Renderer, camera: ICamera, scene: Scene, toDrawingBuffer: boolean, transparentBackground: boolean, postprocessingProps: PostprocessingProps) {
@@ -284,10 +296,7 @@ export class DrawPass {
                 this.colorTarget.bind();
             }
 
-            const outlineEnabled = PostprocessingPass.isEnabled(postprocessingProps) && PostprocessingPass.isTransparentOutlineEnabled(postprocessingProps);
-            const dofEnabled = DofPass.isEnabled(postprocessingProps);
-
-            if (outlineEnabled || dofEnabled) {
+            if (PostprocessingPass.isTransparentDepthRequired(scene, postprocessingProps)) {
                 this.depthTargetTransparent.bind();
                 renderer.clearDepth(true);
                 if (scene.opacityAverage < 1) {
@@ -295,14 +304,39 @@ export class DrawPass {
                 }
             }
 
-            if (PostprocessingPass.isEnabled(postprocessingProps)) {
+            // render transparent primitives
+            const isPostprocessingEnabled = PostprocessingPass.isEnabled(postprocessingProps);
+            if (scene.opacityAverage < 1) {
+                if (isPostprocessingEnabled) {
+                    this.transparentColorTarget.bind();
+                    renderer.clear(false, false, true);
+
+                    if (!this.packedDepth) {
+                        this.depthTextureOpaque.attachFramebuffer(this.transparentColorTarget.framebuffer, 'depth');
+                    } else {
+                        this.colorTarget.depthRenderbuffer?.detachFramebuffer(this.transparentColorTarget.framebuffer);
+                    }
+                }
+
+                renderer.renderBlendedTransparent(scene.primitives, camera);
+
+                if (isPostprocessingEnabled) {
+                    if (!this.packedDepth) {
+                        this.depthTextureOpaque.detachFramebuffer(this.transparentColorTarget.framebuffer, 'depth');
+                    } else {
+                        this.colorTarget.depthRenderbuffer?.detachFramebuffer(this.transparentColorTarget.framebuffer);
+                    }
+                }
+            }
+
+            if (isPostprocessingEnabled) {
                 if (!this.packedDepth) {
                     this.depthTextureOpaque.detachFramebuffer(this.postprocessing.target.framebuffer, 'depth');
                 } else {
                     this.colorTarget.depthRenderbuffer?.detachFramebuffer(this.postprocessing.target.framebuffer);
                 }
 
-                this.postprocessing.render(camera, false, transparentBackground, renderer.props.backgroundColor, postprocessingProps, renderer.light, renderer.ambientColor);
+                this.postprocessing.render(camera, scene, false, transparentBackground, renderer.props.backgroundColor, postprocessingProps, renderer.light, renderer.ambientColor);
 
                 if (!this.packedDepth) {
                     this.depthTextureOpaque.attachFramebuffer(this.postprocessing.target.framebuffer, 'depth');
@@ -331,9 +365,7 @@ export class DrawPass {
                 }
                 target.bind();
             }
-        }
-
-        if (scene.opacityAverage < 1) {
+        } else if (scene.opacityAverage < 1) {
             renderer.renderBlendedTransparent(scene.primitives, camera);
         }
     }