+++ src/mol-plugin-ui/structure/quick-styles.tsx
@@ -75,6 +75,7 @@ export class QuickStyles extends PurePluginUIComponent {
                             samples: 32,
                             resolutionScale: 1,
                             color: Color(0x000000),
+                            transparentThreshold: 0.4,
                         }
                     },
                     shadow: { name: 'off', params: {} },
@@ -114,6 +115,7 @@ export class QuickStyles extends PurePluginUIComponent {
                                 samples: 32,
                                 resolutionScale: 1,
                                 color: Color(0x000000),
+                                transparentThreshold: 0.4,
                             }
                     },
                     shadow: { name: 'off', params: {} },